slightly modify z_size handling
make it _Atomic, ditch the not-obviously-necessary
CAS (we can just set it atomically, we don't care about
its previous value except diagnostically and only weakly)

in zfs_trunc march z_size downwards in lockstep with
ubc_setsize progress, rather than setting it to the final
result all at once (or leaving it until after ubc_setsize
handling has been done), unless we totally clean the range
after end, in which case we can set it down to the end of
end's page.

we're trying to avoid the situation where a vm_map_msync
is called on the file from another thread, dirtying any
blocks above end.

but we're also trying to avoid the situation where another
thread triggers a pagein of a block above end.

so move z_size after we ubc_setsize.