go back to RL_READER (in part) for zfs_read
we'll try to take a shared RL_READER lock on the range
we're interested in; if we fail, it's because someone has
an exclusive lock on the range (or a bigger range
overlapping it, perhaps as big as 0..UINT64_MAX).

someone with an exclusive lock is likely to be changing
the file size or ubc size -- extending it, probably, but
it could be a truncation or other operation.   so if we
don't get the reader lock, we'll wait to get an exlcusive
lock (RL_WRITER).

before the ubc operations we'll definitely take z_map_lock
so that we don't end up crashing into a deadlock below our
UBC operations (in particular it is hard to suss out
what's happening in that case, whereas waiting on
z_map_lock is noisy).

also, zfs_read is not where we want to be fixing any mismatch
between zp->z_size and ubc size; however, if we hold the
exclusive RL_WRITER we will adjust ubc size to hold the
whole UPL (otherwise we'll complain if the ubc size is not
big enough for the whole UPL).  (this is probably noisy
overkill, though)