give more detail on the ERROR 5 precious problem
so pageoutv2_helper got the locks itself, presumably this means it was
called by the vm system itself or by userland msync(2).  report whether
the file was mmapped/mmaped-writeable and whether the UPL_MSYNC flag was
given to zfs_vnop_pageoutv2.

perhaps should range lock the whole 0 -> UINT64_MAX for the file in a
distinguishing case like this.