ubc to uio: WARNING case -> return 0
the page is dirty, which is not great, but maybe returning 0 will let it
be cleaned up by someone else.

of course it might just loop around forever if the reader keeps
hammering away, so this might end up becoming an error too.

(it certainly appears to result in a stuckage below m_o_l_r)