UPL_COMMIT_INACTIVATE on tails and retired precious pages
This clears the reference bit that got turned on when we touched the
mapping (cf buf_unmap(), vfs_bio.c etc.).

This is used for B_AGE in cluster_iodone() too, which is essentially the
behaviour we want: make these pages eligible for early removal by the VM
system under load.