experiment: mapped writeable? return pages unchanged
If the file is mapped writeable then we could be at the end of the for
loop in pageoutv2 racing a page-dirtying that our ubc_upl_commit_range
would erase.   Instead, abort the UPL, returning the pages to the cache
unchanged (except that when we made the UPL we did disturb the HW dirty
bits).

vnode_pageout vfs_cluster do something similar

We have a trade-off here between losing updates arriving while we're in
pageout (which can be triggered in a variety of ways) and not cleaning
a file's dirty pages out of the cache, even if they are not mapped.

This is unlikely to pose a real-world problem (whereas the userland
page-dirtying vs pageoutv2 *is* a real world problem, since our dmu
activity on dirty pages can take some time and since we can get here
without any involvement of the userland activity doing the
page-dirtying).  However, we could commit anyway under certain
conditions, such as if we can distinguish between a userland msync and
other paths to the commit/abort decision (e.g. a system reap, a vfs
sync, an action on the file by a process not involved in the mmap), if
system memory is tight, if the file is mapped writable for too long,
etc.

For now count bleat, and if bleating is frequent then dprintf the bleat
and think about the previous paragraph.