dump read-only filesystem/immutable-file pages earlier
if we unmount a snapshot, we may get called by the
ubc_msync that VFS does; even if there are no dirty pages
we would do the safe-to-dmu_write dance of extending a
recordsize.  As we do page rounding, a small file would
be at risk of having its blksz bumped going down this
path.   For the case where we actually write out pages,
that's completely fine.  However, since we won't write
out pages in the read-only/immutable case, we shouldn't
be extending the file.

Slide the read-only tests to before the safe-to-dmu_write
code (and indeed before range locking), and if we are r/o
or immutable dump the pages and exit_abort.

DRY suggests there should be an exit_dump_pages_abort goto
target for these and the already-existing similar case.
Next step. :-)