don't clean precious pages in mappedread
(also possibly shouldn't clean at all for ROFS, ROFILE or
immutable file)

the issue here is that VNOP_PAGEIN zeroing out the tail of
the last page of a file creates a precious page; we can
certainly clean this, but we generate a lot of logspam if
zfs_msync->pageoutv2 results in an EROFS or similar, even
though the precious page is just ditched as expected.

Moreover, when there is lots of contention on the last
page, which can happen to executables for instance, the
cleaning of the precious page results in looping in
mappedread *and* in pageoutv2, delaying execution of the
programme.  This is especially visible when gsed is on a
r/o fs and a -j 8 make is done for zfs.