pageout: explicitly abort mmapped pages
rather than skipping over pages and letting the final
disposition (commit or abort) deal with them, if our
file is mmapped, we explicitly abort any
absent/non-present pages.

this *shouldn't* make a difference, but CoreData appears
to get upset when the sqlite -shm file is paged out under
heavy memory load.

(it may be on the pagein side, stand by for next commit)