XXX: (temporary) disable LDI for Trinity pool
Trinity is waiting on a replacement physdisk which when
scrubbed reliably produces short I/Os (typically a write)
which under LDI terminates the scrub and begins a
resilver.  It ONLY happens when scrubbing, and only when
LDI is enabled.

(Without LDI the scrub continues and the short I/Os are
retried by IOKit or somewhere, leading to a blinkenlight
stall on the scrub, which eventually recovers and
continues.)