direct evictions from dbuf cache should be gentler on CPU
a lot of concurrent zfs lists on a low-latency
pool (e.g. SSD or RAM or spinnies with L2 cache) will
steal all the CPU resources from userland.   znapzend in
particular does this when the number of datasets to be
backed up in a run is large.

track directly evicting threads with an atomic counter; if
there's more than one then before doing the eviction, use
IODelay(1) which does a PAUSE / REP NOP for a microsecond,
allowing progress on the other hyperthread; if there's
lots (max_ncpus / 2) then use IOSleep(1), which
deschedules the thread for a millisecond.